refactor: change page structure
