+++ index.html
@@ -3,25 +3,11 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Posts</title>
-    <link rel="stylesheet" href="/src/style.css" />
+    <title>Social Grid</title>
   </head>
   <body>
-    <h1>Posts</h1>
-
-    <input type="text" id="search" placeholder="Search by keyword..." />
-    <table>
-      <thead>
-        <tr>
-          <th>ID</th>
-          <th>Title</th>
-          <th>Body</th>
-          <th></th>
-        </tr>
-      </thead>
-      <tbody id="post-container"></tbody>
-    </table>
-
-    <script src="main.ts" type="module"></script>
+    <script>
+      window.location.replace('/posts/');
+    </script>
   </body>
 </html>
