+++ report/index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Report</title>
+    <link rel="stylesheet" href="style.css" />
   </head>
   <body>
     <h1>Report</h1>
@@ -26,6 +27,6 @@
       </table>
     </section>
 
-    <script type="module" src="/report/report.ts"></script>
+    <script type="module" src="/report/index.ts"></script>
   </body>
 </html>
