+++ report/index.ts
@@ -1,0 +1,3 @@
+import { init } from './report';
+
+init();
